Divide edit message tests into 2 collections
