+++ test/Telegram.Bot.Tests.Integ/Update Messages/EditMessageContentTests.cs
@@ -1,5 +1,4 @@
 using System;
-using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using Telegram.Bot.Tests.Integ.Framework;
@@ -22,50 +21,6 @@ namespace Telegram.Bot.Tests.Integ.Update_Messages
         public EditMessageContentTests(TestsFixture fixture)
         {
             _fixture = fixture;
-        }
-
-        [OrderedFact(DisplayName = FactTitles.ShouldEditMessageText)]
-        [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SendMessage)]
-        [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.EditMessageText)]
-        public async Task Should_Edit_Message_Text()
-        {
-            await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldEditMessageText);
-
-            const string originalMessagePrefix = "original\n";
-            (MessageEntityType Type, string Value)[] entityValueMappings = {
-                (MessageEntityType.Bold, "<b>bold</b>"),
-                (MessageEntityType.Italic, "<i>italic</i>"),
-            };
-            string messageText = originalMessagePrefix +
-                    string.Join("\n", entityValueMappings.Select(tuple => tuple.Value));
-
-            Message originalMessage = await BotClient.SendTextMessageAsync(
-                chatId: _fixture.SupergroupChat.Id,
-                text: messageText,
-                parseMode: ParseMode.Html
-            );
-
-            DateTime timeBeforeEdition = DateTime.UtcNow;
-            await Task.Delay(1_000);
-
-            const string modifiedMessagePrefix = "modified\n";
-            messageText = modifiedMessagePrefix +
-                    string.Join("\n", entityValueMappings.Select(tuple => tuple.Value));
-            Message editedMessage = await BotClient.EditMessageTextAsync(
-                chatId: originalMessage.Chat.Id,
-                messageId: originalMessage.MessageId,
-                text: messageText,
-                parseMode: ParseMode.Html
-            );
-
-            Assert.StartsWith(modifiedMessagePrefix, editedMessage.Text);
-            Assert.Equal(originalMessage.MessageId, editedMessage.MessageId);
-            Assert.True(timeBeforeEdition < editedMessage.EditDate);
-
-            Assert.Equal(
-                entityValueMappings.Select(tuple => tuple.Type),
-                editedMessage.Entities.Select(e => e.Type)
-            );
         }
 
         [OrderedFact(DisplayName = FactTitles.ShouldEditInlineMessageText)]
@@ -123,33 +78,6 @@ namespace Telegram.Bot.Tests.Integ.Update_Messages
             );
         }
 
-        [OrderedFact(DisplayName = FactTitles.ShouldEditMessageMarkup)]
-        [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SendMessage)]
-        [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.EditMessageReplyMarkup)]
-        public async Task Should_Edit_Message_Markup()
-        {
-            await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldEditMessageMarkup);
-
-            Message message = await BotClient.SendTextMessageAsync(
-                chatId: _fixture.SupergroupChat.Id,
-                text: "Inline keyboard will be updated shortly",
-                replyMarkup: (InlineKeyboardMarkup) "Original markup"
-            );
-
-            DateTime timeBeforeEdition = DateTime.UtcNow;
-            await Task.Delay(1_000);
-
-            Message editedMessage = await BotClient.EditMessageReplyMarkupAsync(
-                chatId: message.Chat.Id,
-                messageId: message.MessageId,
-                replyMarkup: "Edited 👍"
-            );
-
-            Assert.Equal(message.MessageId, editedMessage.MessageId);
-            Assert.Equal(message.Text, editedMessage.Text);
-            Assert.True(timeBeforeEdition < editedMessage.EditDate);
-        }
-
         [OrderedFact(DisplayName = FactTitles.ShouldEditInlineMessageMarkup)]
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.AnswerInlineQuery)]
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.EditMessageReplyMarkup)]
@@ -194,39 +122,6 @@ namespace Telegram.Bot.Tests.Integ.Update_Messages
                 inlineMessageId: callbackQUpdate.CallbackQuery.InlineMessageId,
                 replyMarkup: "✌ Edited 👌"
             );
-        }
-
-        [OrderedFact(DisplayName = FactTitles.ShouldEditMessageCaption)]
-        [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SendPhoto)]
-        [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.EditMessageCaption)]
-        public async Task Should_Edit_Message_Caption()
-        {
-            await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldEditMessageCaption);
-
-            Message originalMessage;
-            using (Stream stream = System.IO.File.OpenRead(Constants.FileNames.Photos.Bot))
-            {
-                originalMessage = await BotClient.SendPhotoAsync(
-                    chatId: _fixture.SupergroupChat.Id,
-                    photo: stream,
-                    caption: "Message caption will be updated shortly"
-                );
-            }
-
-            DateTime timeBeforeEdition = DateTime.UtcNow;
-            await Task.Delay(1_000);
-
-            const string newCaption = "Caption is edited.";
-
-            Message editedMessage = await BotClient.EditMessageCaptionAsync(
-                chatId: originalMessage.Chat.Id,
-                messageId: originalMessage.MessageId,
-                caption: newCaption
-            );
-
-            Assert.Equal(originalMessage.MessageId, editedMessage.MessageId);
-            Assert.Equal(newCaption, editedMessage.Caption);
-            Assert.True(timeBeforeEdition < editedMessage.EditDate);
         }
 
         [OrderedFact(DisplayName = FactTitles.ShouldEditInlineMessageCaption)]
@@ -276,15 +171,9 @@ namespace Telegram.Bot.Tests.Integ.Update_Messages
 
         private static class FactTitles
         {
-            public const string ShouldEditMessageText = "Should edit a message's text";
-
             public const string ShouldEditInlineMessageText = "Should edit an inline message's text";
 
-            public const string ShouldEditMessageMarkup = "Should edit a message's markup";
-
             public const string ShouldEditInlineMessageMarkup = "Should edit an inline message's markup";
-
-            public const string ShouldEditMessageCaption = "Should edit a message's caption";
 
             public const string ShouldEditInlineMessageCaption = "Should edit an inline message's caption";
         }
